add huawei to schemagen
