+++ bdap-mgr/src/test/resources/testFlow.properties
@@ -1,8 +1,8 @@
 jdk.bin=C:\\dev\\Java\\jdk1.8.0_60\\bin
 tmp.dir=C:\\tmp
 platform.local.dist=C:\\mydoc\\myprojects\\bdap\\bdap-VVERSIONN\\engine
-platform.remote.lib=/bdap-r0.2.0/engine/lib
-platform.coordinate.xml=/bdap-r0.2.0/cfg/coordinator.xml
+platform.remote.lib=/bdap-VVERSIONN/engine/lib
+platform.coordinate.xml=/bdap-VVERSIONN/cfg/coordinator.xml
 projects=project1
 project1.local.dir=C:\\mydoc\\myprojects\\bdap\\bdap-mgr\\src\\test\\resources\\
 project1.hdfs.dir=/project1/
