+++ bdap-schemagen/pom.xml
@@ -99,7 +99,7 @@
 							<transformers>
 								<transformer
 									implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-									<mainClass>bdap.schemagen.tools.SchemaGenerator</mainClass>
+									<mainClass>bdap.schemagen.femto.SchemaGenerator</mainClass>
 								</transformer>
 							</transformers>
 						</configuration>
